(SERVER-1408) Modify poms to support early release of jars
This commit modifies some pom files to make it possible to release
the `jruby-core` and `jruby-stdlib` jars independently of the
upstream release cycle.  This is necessary so that we can
ship some critical bug fixes that have been merged upstream
but have not yet been released.

The commit also adds a README-PUPPETLABS.md file explaining the
changes and how to do the releases.